@@ -22,58 +22,58 @@
   "dependencies": {
     "@elastic/elasticsearch": "~7.13.0",
     "@nestjs/axios": "^0.0.3",
-    "@nestjs/common": "^8.2.1",
-    "@nestjs/core": "^8.2.1",
+    "@nestjs/common": "^8.2.3",
+    "@nestjs/core": "^8.2.3",
     "@nestjs/elasticsearch": "^8.0.0",
     "@nestjs/jwt": "^8.0.0",
     "@nestjs/mongoose": "^9.0.1",
     "@nestjs/passport": "^8.0.1",
-    "@nestjs/platform-express": "^8.2.1",
+    "@nestjs/platform-express": "^8.2.3",
     "@nestjs/schedule": "^1.0.2",
-    "@nestjs/swagger": "^5.1.4",
+    "@nestjs/swagger": "^5.1.5",
     "@types/dockerode": "^3.3.0",
     "@types/passport-jwt": "^3.0.6",
     "axios": "^0.24.0",
     "bson-objectid": "^2.0.1",
     "class-transformer": "^0.4.0",
-    "class-validator": "^0.13.1",
+    "class-validator": "^0.13.2",
     "dockerode": "^3.3.1",
     "glob-to-regexp": "^0.4.1",
     "gunzip-maybe": "^1.4.2",
-    "mongoose": "^6.0.13",
+    "mongoose": "^6.1.2",
     "passport": "^0.5.0",
     "passport-jwt": "^4.0.0",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
     "rxjs": "^7.4.0",
-    "swagger-ui-express": "^4.1.6",
+    "swagger-ui-express": "^4.2.0",
     "tar-stream": "^2.2.0"
   },
   "devDependencies": {
-    "@nestjs/cli": "^8.1.4",
+    "@nestjs/cli": "^8.1.6",
     "@nestjs/schematics": "^8.0.5",
-    "@nestjs/testing": "^8.2.1",
+    "@nestjs/testing": "^8.2.3",
     "@types/cron": "^1.7.3",
     "@types/express": "^4.17.13",
     "@types/glob-to-regexp": "^0.4.1",
     "@types/gunzip-maybe": "^1.4.0",
-    "@types/jest": "^27.0.2",
-    "@types/node": "^16.11.7",
+    "@types/jest": "^27.0.3",
+    "@types/node": "^16.11.13",
     "@types/supertest": "^2.0.11",
     "@types/tar-stream": "^2.2.2",
-    "@typescript-eslint/eslint-plugin": "^5.4.0",
-    "@typescript-eslint/parser": "^5.4.0",
-    "eslint": "^8.2.0",
+    "@typescript-eslint/eslint-plugin": "^5.7.0",
+    "@typescript-eslint/parser": "^5.7.0",
+    "eslint": "^8.4.1",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "jest": "^27.3.1",
-    "prettier": "^2.4.1",
+    "jest": "^27.4.5",
+    "prettier": "^2.5.1",
     "supertest": "^6.1.6",
-    "ts-jest": "^27.0.7",
+    "ts-jest": "^27.1.1",
     "ts-loader": "^9.2.6",
     "ts-node": "^10.4.0",
-    "tsconfig-paths": "^3.11.0",
-    "typescript": "^4.4.4"
+    "tsconfig-paths": "^3.12.0",
+    "typescript": "^4.5.4"
   },
   "jest": {
     "moduleFileExtensions": [
